==> 07a-lda-kfold-xval.Rmd <==
---
title: "LDA K-fold cross-validation"
output: html_notebook
---

# Goals

  * Find the optimal number of topics using a cross-validation approach
  * Select a final LDA model using the aggressive TeX removal sparse matrix

# Libraries used

```{r, warning=F, message=F, error=F, comment=F}
library(dplyr)
library(wranglR)
library(ggplot2)
library(stringr)
library(tm)
library(slam)
library(topicmodels)
```

# Load the data

See [06-dataset-comparison.Rmd](https://github.com/phively/uchicago-thesis/blob/master/06-dataset-comparison.Rmd) for details on how the sparse matrix was generated.

```{r, message=F}
# Load sparse matrix with no tex
load(file = "data/sparse_matrix_no_tex.zip")
```

# Cross-validation setup

Set the number of groups $K$ and generate indices.

```{r}
# Set K
K <- 5

# Create k groups, as close to equal size as possible
set.seed(93401)
rand_ind <- KFoldXVal(1:sparse$nrow, k = K)
```

Create the functions to be used for cross-validation.

```{r}
# Cross-validation matrix generator function, split_set()
source("R/function-split_set.R")

# Cross-validation iterator
xval <- function(dtm, index, ntopics, seed, ...) {
  # Result models list
  models <- list()
  
  # Cross-validation loop
  for (i in 1:K) {
    # Create dataset for fitting
    this_ind <- unlist(index)[unlist(index) %nin% index[[i]]]
    this_data <- split_set(dtm, this_ind)
    
    # Create a topic model
    models[[i]] <- LDA(this_data, k = ntopics, method = "VEM", control = list(seed = seed, ...))
  }
  
  return(models)
}
```

# Topic modeling

## Modeling step

### 5 topics

```{r}
now <- Sys.time()
models5 <- xval(dtm = sparse, index = rand_ind, ntopics = 5, seed = 74749, keep = 1, verbose = 1)
save(models5, file = "results/lda_k_5.zip", compress = "xz")
(t5 <- Sys.time() - now)
```

### 10 topics

```{r}
now <- Sys.time()
models10 <- xval(dtm = sparse, index = rand_ind, ntopics = 10, seed = 90672, keep = 1, verbose = 1)
save(models10, file = "results/lda_k_10.zip", compress = "xz")
(t10 <- Sys.time() - now)
```

### 15 topics

```{r}
now <- Sys.time()
models15 <- xval(dtm = sparse, index = rand_ind, ntopics = 15, seed = 53177, keep = 1, verbose = 1)
save(models15, file = "results/lda_k_15.zip", compress = "xz")
(t15 <- Sys.time() - now)
```

This takes a really long time. Makes more sense to do the rest on AWS in parallel for pennies an hour....

# Perplexity

## In-sample results

```{r}
# Calculate in-sample perplexity
perplexity <- data.frame(
  model = c("5", "10", "15") %>% rep(each = K),
  in_sample = c(
    lapply(models5, perplexity) %>% unlist(),
    lapply(models10, perplexity) %>% unlist(),
    lapply(models15, perplexity) %>% unlist()
  )
)
```

## Out-of-sample results

```{r}
# Calculate out-of-sample perplexity
perp5 <- rep(0, times = K)
perp10 <- rep(0, times = K)
perp15 <- rep(0, times = K)
for(i in 1:K) {
  perp5[i] <- perplexity(models5[[i]], newdata = split_set(sparse, rand_ind[[i]]))
  perp10[i] <- perplexity(models10[[i]], newdata = split_set(sparse, rand_ind[[i]]))
  perp15[i] <- perplexity(models15[[i]], newdata = split_set(sparse, rand_ind[[i]]))
}

# Append to perplexity data frame
perplexity <- perplexity %>% mutate(
  out_sample = c(perp5, perp10, perp15)
)

# Save results
write.csv(perplexity, file = "results/lda_k_perp_5_15.csv", row.names = FALSE)
```
