==> 08a-ctm-exploration.Rmd <==
---
title: "Correlated topic models exploration"
output: html_notebook
---

# Goals

  * Explore correlated topic models as implemented by the `stm` package.

# Libraries used

```{r, warning=F, message=F, error=F, comment=F}
library(dplyr)
library(slam)
library(stm)
library(foreach)
library(doParallel)
```

Set up parallel runs

```{r}
registerDoParallel(5)
getDoParWorkers()
```

# Test run of stm

## Load test data

```{r}
# Test data
load(file = "data/sparse_test_smaller.zip")
extra_sparse <- readCorpus(extra_sparse, type = "slam")
```

If covariates are not included, stm acts like a correlated topic model.

## 5-topic test model

```{r}
now <- Sys.time()
m5 <- stm(extra_sparse$documents, extra_sparse$vocab, K = 5, init.type = "Spectral", ngroups = 5)
(t5 <- Sys.time() - now)
```

## 10-topic test model

```{r}
now <- Sys.time()
m10 <- stm(extra_sparse$documents, extra_sparse$vocab, K = 10, init.type = "Spectral", ngroups = 5)
(t10 <- Sys.time() - now)
```

## Explore test results

```{r}
labelTopics(m5)
topicCorr(m5)$poscor %>% data.frame() %>% signif(2)
topicQuality(m5, documents = extra_sparse$documents)
plot(m5)
plot(topicCorr(m5))
```

```{r}
labelTopics(m10)
topicCorr(m10)$poscor %>% data.frame() %>% signif(2)
topicQuality(m10, documents = extra_sparse$documents)
plot(m10)
plot(topicCorr(m10))
```

## Try the K-searching method

```{r}
now <- Sys.time()
manyK <- foreach(this.k = seq(3, 15, by = 3), .packages = "stm") %dopar% {
  output <- searchK(extra_sparse$documents, extra_sparse$vocab, K = this.k, proportion = .5, heldout.seed = 123, verbose = FALSE)
  return(output$results)
}
(tK <- Sys.time() - now)
```

## searchK results

```{r, fig.height=8}
for.plotting <- list(results = bind_rows(manyK), call = NULL)
attr(for.plotting, "class") <- "searchK"
plot(for.plotting)
```
