==> 08b-ctm-no-diag.Rmd <==
---
title: "Correlated topic model with diagonal covariance"
output: html_notebook
---

# Goal

  * Use the `stm` package to fit CTMs with diagonal $\Sigma$
  * Have a baseline for comparing structural topic models

# Libraries used

```{r, warning=F, message=F, error=F, comment=F}
library(tidyverse)
library(slam)
library(stm)
library(foreach)
library(doParallel)
```

Set up parallel runs

```{r}
registerDoParallel(5)
getDoParWorkers()
```

# Load the final data

See [06-dataset-comparison.Rmd](https://github.com/phively/uchicago-thesis/blob/master/06-dataset-comparison.Rmd) for details on how the sparse matrix was generated.

```{r, message=F}
# Load sparse matrix with no tex
load(file = "data/sparse_matrix_no_tex.zip")
sparse <- readCorpus(sparse, type = "slam")
```

# CTM method without correlations

Topics $K \in \left\{ 5, 10, \dots, 50 \right\}$

```{r}
# 5
gc()
nc_ctm05 <- searchK(sparse$documents, sparse$vocab, K = 5, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm05, file = "results/nc_ctm05.zip", compress = "xz", compression_level = 1)

# 10
gc()
nc_ctm10 <- searchK(sparse$documents, sparse$vocab, K = 10, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm10, file = "results/nc_ctm10.zip", compress = "xz", compression_level = 1)

# 15
gc()
nc_ctm15 <- searchK(sparse$documents, sparse$vocab, K = 15, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm15, file = "results/nc_ctm15.zip", compress = "xz", compression_level = 1)

# 20
gc()
nc_ctm20 <- searchK(sparse$documents, sparse$vocab, K = 20, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm20, file = "results/nc_ctm20.zip", compress = "xz", compression_level = 1)

# 25
gc()
nc_ctm25 <- searchK(sparse$documents, sparse$vocab, K = 25, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm25, file = "results/nc_ctm25.zip", compress = "xz", compression_level = 1)

# 30
gc()
nc_ctm30 <- searchK(sparse$documents, sparse$vocab, K = 30, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm30, file = "results/nc_ctm30.zip", compress = "xz", compression_level = 1)

# 35
gc()
nc_ctm35 <- searchK(sparse$documents, sparse$vocab, K = 35, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm35, file = "results/nc_ctm35.zip", compress = "xz", compression_level = 1)

# 40
gc()
nc_ctm40 <- searchK(sparse$documents, sparse$vocab, K = 40, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm40, file = "results/nc_ctm40.zip", compress = "xz", compression_level = 1)

# 45
gc()
nc_ctm45 <- searchK(sparse$documents, sparse$vocab, K = 45, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm45, file = "results/nc_ctm45.zip", compress = "xz", compression_level = 1)

# 50
gc()
nc_ctm50 <- searchK(sparse$documents, sparse$vocab, K = 50, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm50, file = "results/nc_ctm50.zip", compress = "xz", compression_level = 1)
```

```{r}
xval <- bind_rows(nc_ctm05$results, nc_ctm10$results, nc_ctm15$results, nc_ctm20$results, nc_ctm25$results,
  nc_ctm30$results, nc_ctm35$results, nc_ctm40$results, nc_ctm45$results, nc_ctm50$results)
write.csv(xval, file = "results/nc_ctm_xval.csv", row.names = FALSE)
```

```{r}
xval %>% ggplot(aes(x = K, y = heldout)) + geom_point()
```

There's clearly room for improvement.

# CTM method, diagonal $\Sigma$, 60 through 150

```{r}
# 60
gc()
nc_ctm60 <- searchK(sparse$documents, sparse$vocab, K = 60, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm60, file = "results/nc_ctm60.zip", compress = "xz", compression_level = 1)

# 70
gc()
nc_ctm70 <- searchK(sparse$documents, sparse$vocab, K = 70, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm70, file = "results/nc_ctm70.zip", compress = "xz", compression_level = 1)

# 80
gc()
nc_ctm80 <- searchK(sparse$documents, sparse$vocab, K = 80, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm80, file = "results/nc_ctm80.zip", compress = "xz", compression_level = 1)

# 90
gc()
nc_ctm90 <- searchK(sparse$documents, sparse$vocab, K = 90, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm90, file = "results/nc_ctm90.zip", compress = "xz", compression_level = 1)

# 100
gc()
nc_ctm100 <- searchK(sparse$documents, sparse$vocab, K = 100, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm100, file = "results/nc_ctm100.zip", compress = "xz", compression_level = 1)

# 110
gc()
nc_ctm110 <- searchK(sparse$documents, sparse$vocab, K = 110, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm110, file = "results/nc_ctm110.zip", compress = "xz", compression_level = 1)

# 120
gc()
nc_ctm120 <- searchK(sparse$documents, sparse$vocab, K = 120, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm120, file = "results/nc_ctm120.zip", compress = "xz", compression_level = 1)

# 130
gc()
nc_ctm130 <- searchK(sparse$documents, sparse$vocab, K = 130, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm130, file = "results/nc_ctm130.zip", compress = "xz", compression_level = 1)

# 140
gc()
nc_ctm140 <- searchK(sparse$documents, sparse$vocab, K = 140, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm140, file = "results/nc_ctm140.zip", compress = "xz", compression_level = 1)

# 150
gc()
nc_ctm150 <- searchK(sparse$documents, sparse$vocab, K = 150, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), proportion = .5, heldout.seed = 13316, sigma.prior = 1)
save(nc_ctm150, file = "results/nc_ctm150.zip", compress = "xz", compression_level = 1)
```

# Update cross-validation data frame

```{r}
xval <- bind_rows(xval,
  nc_ctm60$results, nc_ctm70$results, nc_ctm80$results, nc_ctm90$results, nc_ctm100$results,
  nc_ctm110$results, nc_ctm120$results, nc_ctm130$results, nc_ctm140$results, nc_ctm150$results)
write.csv(xval, file = "results/nc_ctm_xval.csv", row.names = FALSE)
```

```{r}
xval %>% ggplot(aes(x = K, y = heldout)) + geom_point()
```

# New strategy -- true hold-out cross-validation

## Holdout set

```{r}
# Create holdout sets
N <- length(sparse$documents)
holdout1 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = 1, seed = 83319)
holdout2 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = 1, seed = 5052)
holdout3 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = 1, seed = 54551)
holdout4 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = 1, seed = 19245)
holdout5 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = 1, seed = 56423)

holdout <- list(holdout1, holdout2, holdout3, holdout4, holdout5)
```

# Parallelization

```{r}
# Parallel function
xval <- function(k) {
  foreach(i = 1:5) %dopar% {
    stm(holdout[[i]]$documents, holdout[[i]]$vocab, K = k, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout[[i]]$vocab)), sigma.prior = 1)
  }
}
```

# K = 5 (test)

```{r}
# 5
nc_5 <- xval(k = 5)
save(nc_5, file = "results/nc_5.zip", compress = "xz", compression_level = 1)
```

## K = 10 to 60

```{r}
# 10
gc()
nc_10 <- xval(k = 10)
save(nc_10, file = "results/nc_10.zip", compress = "xz", compression_level = 1)

# 15
gc()
nc_15 <- xval(k = 15)
save(nc_15, file = "results/nc_15.zip", compress = "xz", compression_level = 1)

# 20
gc()
nc_20 <- xval(k = 20)
save(nc_20, file = "results/nc_20.zip", compress = "xz", compression_level = 1)

# 25
gc()
nc_25 <- xval(k = 25)
save(nc_25, file = "results/nc_25.zip", compress = "xz", compression_level = 1)

# 30
gc()
nc_30 <- xval(k = 30)
save(nc_30, file = "results/nc_30.zip", compress = "xz", compression_level = 1)

# 35
gc()
nc_35 <- xval(k = 35)
save(nc_35, file = "results/nc_35.zip", compress = "xz", compression_level = 1)

# 40
gc()
nc_40 <- xval(k = 40)
save(nc_40, file = "results/nc_40.zip", compress = "xz", compression_level = 1)

# 45
gc()
nc_45 <- xval(k = 45)
save(nc_45, file = "results/nc_45.zip", compress = "xz", compression_level = 1)

# 50
gc()
nc_50 <- xval(k = 50)
save(nc_50, file = "results/nc_50.zip", compress = "xz", compression_level = 1)

# 55
gc()
nc_55 <- xval(k = 55)
save(nc_55, file = "results/nc_55.zip", compress = "xz", compression_level = 1)

# 60
gc()
nc_60 <- xval(k = 60)
save(nc_60, file = "results/nc_60.zip", compress = "xz", compression_level = 1)
```

# Evaluate holdout samples

```{r}
test <- list()
for(i in 1:5) {
  test[[i]] <- eval.heldout(nc_5[[i]], holdout[[i]]$missing)$doc.heldout
  test[[5+i]] <- eval.heldout(nc_10[[i]], holdout[[i]]$missing)$doc.heldout
  test[[10+i]] <- eval.heldout(nc_15[[i]], holdout[[i]]$missing)$doc.heldout
  test[[15+i]] <- eval.heldout(nc_20[[i]], holdout[[i]]$missing)$doc.heldout
  test[[20+i]] <- eval.heldout(nc_25[[i]], holdout[[i]]$missing)$doc.heldout
  test[[25+i]] <- eval.heldout(nc_30[[i]], holdout[[i]]$missing)$doc.heldout
  test[[30+i]] <- eval.heldout(nc_35[[i]], holdout[[i]]$missing)$doc.heldout
  test[[35+i]] <- eval.heldout(nc_40[[i]], holdout[[i]]$missing)$doc.heldout
  test[[40+i]] <- eval.heldout(nc_45[[i]], holdout[[i]]$missing)$doc.heldout
  test[[45+i]] <- eval.heldout(nc_50[[i]], holdout[[i]]$missing)$doc.heldout
  test[[50+i]] <- eval.heldout(nc_55[[i]], holdout[[i]]$missing)$doc.heldout
  test[[55+i]] <- eval.heldout(nc_60[[i]], holdout[[i]]$missing)$doc.heldout
}

n <- length(test[[1]]) + length(test[[2]]) + length(test[[3]]) + length(test[[4]]) + length(test[[5]])
# Per-document log likelihood
nc_heldout_lik <- unlist(test) %>% data.frame(
  K = rep(seq(5, 60, by = 5), each = n),
  lik = .
)
write.csv(nc_heldout_lik, file = "results/nc_ctm_heldout_lik.csv", row.names = FALSE)
```

```{r}
nc_heldout_lik %>% filter(K > 5) %>% group_by(K) %>% summarise(mean.lik = mean(lik), sem.lik = sd(lik)/sqrt(n), lci = mean.lik - sem.lik, uci = mean.lik + sem.lik) %>%
  ggplot(aes(x = K, y = mean.lik)) + geom_ribbon(aes(ymin = lci, ymax = uci), alpha = .2) + geom_line(color = "blue") +
  labs(y = "mean per-document holdout likelihood", x = "") + scale_x_continuous(breaks = seq(10, 60, by = 10))
```

Maximum likelihood at $K$ = 25.
