==> 08c-ctm-with-sigma.Rmd <==
---
title: "Correlated topic models K"
output: html_notebook
---

# Goals

  * Identify a suitable number of topics $K$ to use for the CTM.

# Libraries used

```{r, warning=F, message=F, error=F, comment=F}
library(tidyverse)
library(slam)
library(stm)
library(foreach)
library(doParallel)
```

# Load the final data

See [06-dataset-comparison.Rmd](https://github.com/phively/uchicago-thesis/blob/master/06-dataset-comparison.Rmd) for details on how the sparse matrix was generated.

```{r, message=F}
# Load sparse matrix with no tex
load(file = "data/sparse_matrix_no_tex.zip")
sparse <- readCorpus(sparse, type = "slam")
```

# CTM method without correlations

We can get a model like an LDA by enforcing a diagonal $\Sigma$ with `sigma.prior = 1` but note that the topic distribution in this model is not actually a Dirichlet.

```{r}
nc_ctm40 <- stm(sparse$documents, sparse$vocab, K = 40, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)), sigma.prior = 1)
nc_ctm40$time/60
```

# Rerun the 40 topic model as CTM

Correlated topic models relax the $\Sigma$ is diagonal constraint.

```{r}
now <- Sys.time()
ctm40 <- stm(sparse$documents, sparse$vocab, K = 40, init.type = "Spectral", ngroups = 7,
  control = list(maxV = length(sparse$vocab)))
(t40 <- Sys.time() - now)
```

# Compare diagonal and non-diagonal $\Sigma$

```{r, fig.height=8}
# Time to complete
nc_ctm40$time/60
ctm40$time/60
# Correlations
plot(topicCorr(nc_ctm40))
plot(topicCorr(ctm40))
# Most frequent topics
plot(nc_ctm40, n = 5, xlim = c(0, .1))
plot(ctm40, n = 5, xlim = c(0, .1))
```

# Other numbers of topics

That was surprisingly fast. Try every 5 as before.

```{r}
ctm5  <- stm(sparse$documents, sparse$vocab, K = 5,  init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
ctm10 <- stm(sparse$documents, sparse$vocab, K = 10, init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
ctm15 <- stm(sparse$documents, sparse$vocab, K = 15, init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
ctm20 <- stm(sparse$documents, sparse$vocab, K = 20, init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
ctm25 <- stm(sparse$documents, sparse$vocab, K = 25, init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
ctm30 <- stm(sparse$documents, sparse$vocab, K = 30, init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
ctm35 <- stm(sparse$documents, sparse$vocab, K = 35, init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
ctm45 <- stm(sparse$documents, sparse$vocab, K = 45, init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
ctm50 <- stm(sparse$documents, sparse$vocab, K = 50, init.type = "Spectral", ngroups = 7, control = list(maxV = length(sparse$vocab)))
```

# Multiple K

```{r}
now <- Sys.time()
manyK <- foreach(this.k = seq(50, 5, by = -5), .packages = "stm") %dopar% {
  output <- searchK(sparse$documents, sparse$vocab, K = this.k, init.type = "Spectral", ngroups = 7,
  proportion = .5, heldout.seed = 13316, verbose = FALSE)
  return(output$results)
}
(tK <- Sys.time() - now)
```

## searchK results

```{r, fig.height=8}
manyK <- list(results = bind_rows(manyK), call = NULL)
attr(manyK, "class") <- "searchK"
plot(manyK)
```

## Try additional K

```{r}
now <- Sys.time()
moreK <- foreach(this.k = seq(100, 55, by = -5), .packages = "stm") %dopar% {
  output <- searchK(sparse$documents, sparse$vocab, K = this.k, init.type = "Spectral", ngroups = 7,
  proportion = .5, heldout.seed = 13316, verbose = FALSE)
  return(output$results)
}
(tK2 <- Sys.time() - now)
```

```{r, fig.height=8}
tmp <- manyK
tmp$results <- rbind(bind_rows(moreK), manyK$results)
plot(tmp)
```

# Save results

```{r}
write.csv(rbind(manyK$results, bind_rows(moreK)), file = "results/ctm_holdout.csv")
```

```{r}
save(ctm5, ctm10, ctm15, ctm20, ctm25, ctm30, ctm35, ctm40, ctm45, ctm50, nc_ctm40, manyK, moreK,
     file ="results/ctm_by_5.zip", compress = "xz", compression_level = 1)
```

# Proper hold-out cross-validation

## Create holdout groups

```{r}
# Create holdout sets
N <- length(sparse$documents)
holdout1 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = .5, seed = 63321)
holdout2 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = .5, seed = 13434)
holdout3 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = .5, seed = 81249)

holdout <- list(holdout1, holdout2, holdout3)
```

## K = 5 through 25

```{r}
# 5
ctm_05 <- stm(holdout1$documents, holdout1$vocab, K = 5, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_05, file = "results/ctm_05.zip", compress = "xz", compression_level = 1)

# 10
ctm_10 <- stm(holdout1$documents, holdout1$vocab, K = 10, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_10, file = "results/ctm_10.zip", compress = "xz", compression_level = 1)

# 15
ctm_15 <- stm(holdout1$documents, holdout1$vocab, K = 15, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_15, file = "results/ctm_15.zip", compress = "xz", compression_level = 1)

# 20
ctm_20 <- stm(holdout1$documents, holdout1$vocab, K = 20, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_20, file = "results/ctm_20.zip", compress = "xz", compression_level = 1)

# 25
ctm_25 <- stm(holdout1$documents, holdout1$vocab, K = 25, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_25, file = "results/ctm_25.zip", compress = "xz", compression_level = 1)
```

## K = 30 through 100

```{r}
# 30
ctm_30 <- stm(holdout1$documents, holdout1$vocab, K = 30, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_30, file = "results/ctm_30.zip", compress = "xz", compression_level = 1)

# 40
ctm_40 <- stm(holdout1$documents, holdout1$vocab, K = 40, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_40, file = "results/ctm_40.zip", compress = "xz", compression_level = 1)

# 50
ctm_50 <- stm(holdout1$documents, holdout1$vocab, K = 50, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_50, file = "results/ctm_50.zip", compress = "xz", compression_level = 1)

# 60
ctm_60 <- stm(holdout1$documents, holdout1$vocab, K = 60, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_60, file = "results/ctm_60.zip", compress = "xz", compression_level = 1)

# 70
ctm_70 <- stm(holdout1$documents, holdout1$vocab, K = 70, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_70, file = "results/ctm_70.zip", compress = "xz", compression_level = 1)

# 80
ctm_80 <- stm(holdout1$documents, holdout1$vocab, K = 80, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_80, file = "results/ctm_80.zip", compress = "xz", compression_level = 1)

# 90
ctm_90 <- stm(holdout1$documents, holdout1$vocab, K = 90, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_90, file = "results/ctm_90.zip", compress = "xz", compression_level = 1)

# 100
ctm_100 <- stm(holdout1$documents, holdout1$vocab, K = 100, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_100, file = "results/ctm_100.zip", compress = "xz", compression_level = 1)

# 125
ctm_125 <- stm(holdout1$documents, holdout1$vocab, K = 125, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_125, file = "results/ctm_125.zip", compress = "xz", compression_level = 1)

# 150
ctm_150 <- stm(holdout1$documents, holdout1$vocab, K = 150, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_150, file = "results/ctm_150.zip", compress = "xz", compression_level = 1)

# 175
ctm_175 <- stm(holdout1$documents, holdout1$vocab, K = 175, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_175, file = "results/ctm_175.zip", compress = "xz", compression_level = 1)

# 200
ctm_200 <- stm(holdout1$documents, holdout1$vocab, K = 200, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_200, file = "results/ctm_200.zip", compress = "xz", compression_level = 1)

# 250
ctm_250 <- stm(holdout1$documents, holdout1$vocab, K = 250, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_250, file = "results/ctm_250.zip", compress = "xz", compression_level = 1)

# 500
ctm_500 <- stm(holdout1$documents, holdout1$vocab, K = 500, init.type = "Spectral", ngroups = 5,
      control = list(maxV = length(holdout1$vocab)))
save(ctm_500, file = "results/ctm_500.zip", compress = "xz", compression_level = 1)
```

# Check results

```{r}
test <- c(
  eval.heldout(ctm_05, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_10, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_15, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_20, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_25, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_30, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_40, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_50, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_60, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_70, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_80, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_90, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_100, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_125, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_150, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_175, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_200, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_250, holdout1$missing)$doc.heldout,
  eval.heldout(ctm_500, holdout1$missing)$doc.heldout
)

n <- length(holdout1$missing$index)
# Per-document log likelihood
ctm_heldout_lik <- test %>% data.frame(
  K = rep(seq(5, 25, by = 5), each = n) %>%
    c(rep(seq(30, 100, by = 10), each = n),
      rep(c(125, 150, 175, 200, 250, 500), each = n)),
  lik = .
)

# Save results
write.csv(ctm_heldout_lik, file = "results/ctm_heldout_lik.csv", row.names = FALSE)
```

```{r}
ctm_heldout_lik %>% group_by(K) %>% summarise(mean.lik = mean(lik), sem.lik = sd(lik)/sqrt(n), lci = mean.lik - sem.lik, uci = mean.lik + sem.lik) %>%
  ggplot(aes(x = K, y = mean.lik)) + geom_ribbon(aes(ymin = lci, ymax = uci), alpha = .2) + geom_line(color = "blue") +
  labs(y = "mean per-document holdout log likelihood", x = "")

ctm_heldout_lik %>% ggplot(aes(x = K, y = lik, group = K)) + geom_boxplot(alpha = .25)
```

Held-out log likelihood increases monotonically with $K$ up to 500 topics? I don't buy it.
