==> 09a-stm-metadata-creation.Rmd <==
---
title: "STM metadata"
output: html_notebook
---

# Goal

  * Create metadata to be used for structural topic models
  * Familiarise myself with model specifications used by the `stm` package
  * Experiment with STM visualization techniques
  
# Libraries

```{r}
library(tidyverse)
library(lubridate)
library(stm)
library(foreach)
library(doParallel)
```

# Load posts metadata

```{r}
# Read zipped posts.csv
# Currently includes CreationDate, Score, ViewCount, LastActivityDate, AnswerCount, FavoriteCount, Tags
source("R/read-PostsData-save-metadata.R")
```

# Metadata exploration

```{r}
summary(metadata)
```

There are many NAs under FavoriteCount.

```{r}
metadata$Id[which(is.na(metadata$FavoriteCount)) %>% head()]
```

Looks like it's fine to fill these in with 0.

```{r}
metadata <- metadata %>% mutate(
  FavoriteCount = ifelse(is.na(FavoriteCount), 0, FavoriteCount)
)
summary(metadata$FavoriteCount)
```

Add an index to metadata to map between Id number and document number.

```{r}
metadata <- metadata %>% mutate(
  Doc_Id = 1:length(Id)
)
```

Add a "number of months since beginning of dataset" counter

```{r}
start_dt <- metadata$CreationDate %>% min() %>% floor_date(unit = "month")
metadata <- metadata %>% mutate(
  Months_Since_Feb_09 = (start_dt %--% metadata$CreationDate)/months(1)
)
```

Identify questions tagged as homework, `<self-study>`, or coding in R, `<r>`.

```{r}
metadata <- metadata %>% mutate(
  R_Tag = Tags %>% str_detect('<r>') %>% as.numeric(),
  HW_Tag = Tags %>% str_detect('<self-study>') %>% as.numeric()
)
```

Save the cleaned metadata.

```{r}
save(metadata, file = "data/posts_metadata.zip", compress = "xz")
```

# STM experimentation

## Load data

```{r}
# Load sparse matrix with no tex
load(file = "data/sparse_test_smaller.zip")

# Metadata
small_meta <- metadata %>% head(1000) %>% mutate(
  HW_Tag = factor(HW_Tag, labels = c("no_HW", "HW"))
)

# Put in stm format
extra_sparse <- readCorpus(extra_sparse, type = "slam")
```

## STM with creation date covariate

```{r}
# s() is a B-spline
test_stm10 <- small_meta %>% stm(
  documents = extra_sparse$documents, vocab = extra_sparse$vocab, K = 10,
  init.type = "Spectral", emtol = 1E-3,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag, content = ~ HW_Tag, data = .
)
```

```{r}
# Time in seconds
test_stm10$time/60
```

## Visualize results

```{r, fig.width=8}
# Common topics
plot(test_stm10, n = 5)

# Top documents associated with "Topic 8: matrix, pca, factor"
findThoughts(model = test_stm10, texts = small_meta$CleanBody, n = 2, topics = 8)$docs$`Topic 8` %>% substr(0, 500)
```

## Covariate visualization

```{r}
# Topic labels
topic_label <- apply(cbind(1:10, labelTopics(test_stm10, n = 3)$topics), MARGIN = 1, FUN = paste, collapse = ", ")

# See how the Score covariate affects results
cov_effects <- estimateEffect(~ s(Score) + s(Months_Since_Feb_09) + HW_Tag, test_stm10, metadata = small_meta)
```

```{r, fig.width = 10, fig.height = 6, warning=F, error=F, message=F}
# Mean topic score comparisons
plot(cov_effects, covariate = "Score", method = "difference", cov.value1 = 0, cov.value2 = 5,
     topic = 1:10, labeltype = "custom", custom.labels = topic_label,
     xlab = "Lower score ... Higher score", xlim = c(-.2, .2))
```

```{r}
# Date against proportion
invisible(pdf(file = NULL))
q <- plot(cov_effects, covariate = "Months_Since_Feb_09", topics = 8, method = "continuous")
invisible(dev.off())
q <- data.frame(date = as_datetime(q$x), proportion = q$means[[1]], lci = q$ci[[1]][1, ], uci = q$ci[[1]][2, ])
q %>% ggplot(aes(x = date, y = proportion)) + geom_line() + geom_line(aes(y = lci), linetype = "dotted") +
  geom_line(aes(y = uci), linetype = "dotted") + labs(title = "Topic 8: matrix, pca, factor")

# Continuous score against proportion
pdf(file = NULL)
q <- plot(cov_effects, covariate = "Score", topics = 8, method = "continuous")
dev.off()
q <- data.frame(score = q$x, proportion = q$means[[1]], lci = q$ci[[1]][1, ], uci = q$ci[[1]][2, ])
q %>% ggplot(aes(x = score, y = proportion)) + geom_line() + geom_line(aes(y = lci), linetype = "dotted") +
  geom_line(aes(y = uci), linetype = "dotted") + labs(title = "Topic 8: matrix, pca, factor")

# Confirm that the most popular post is indeed about PCA
small_meta$CleanBody[which(small_meta$Score == max(small_meta$Score))] %>% substr(0, 500)
```

## Covariate effect

```{r}
plot(cov_effects, covariate = "Score", model = test_stm10, method = "continuous",
     moderator = "HW_Tag", moderator.value = "HW", linecol = "red", printlegend = FALSE)
plot(cov_effects, covariate = "Score", model = test_stm10, method = "continuous",
     moderator = "HW_Tag", moderator.value = "no_HW", linecol = "blue", printlegend = FALSE, add = TRUE)
```

No visible difference between homework and non-homework for this small subset.

## Comparing topics

```{r, fig.width=10}
set.seed(123)
plot(test_stm10, type = "perspectives", topics = c(8, 8), topic.names = topic_label,
     main = paste(topic_label[8], topic_label[8], sep = ";  "))
```

```{r, fig.width=6, fig.height=6}
# Topic correlations
plot(topicCorr(test_stm10), vlabels = topic_label)
```

```{r, fig.width=12}
set.seed(123)
plot(test_stm10, type = "perspectives", topics = c(7, 10), topic.names = topic_label,
     main = paste(topic_label[7], topic_label[10], sep = ";  "))
```
