==> 09b-stm-continuous.Rmd <==
---
title: "STM continuous covariates"
output: html_notebook
---

# Goal

  * Try fitting STM dependent on the two continuous covariates I had identified
  * Best cross-validated model
  
# Libraries

```{r}
library(tidyverse)
library(stm)
```

# Load posts and metadata

```{r}
# Load sparse matrix with no tex
load(file = "data/sparse_matrix_no_tex.zip")
sparse <- readCorpus(sparse, type = "slam")

# Load metadata saved in (09a)
load("data/posts_metadata.zip")
metadata <- metadata %>% mutate(
  R_Tag = factor(R_Tag, labels = c("no_R", "R")),
  HW_Tag = factor(HW_Tag, labels = c("no_HW", "HW"))
)
```

# Holdout set

```{r}
# Create holdout sets
N <- length(sparse$documents)
holdout1 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = .95, seed = 77354)
holdout2 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = .95, seed = 72902)
holdout3 <- make.heldout(documents = sparse$documents, vocab = sparse$vocab,
  N = N/5, proportion = .95, seed = 95180)
```

# STM cross-validation with metadata

## First holdout set

```{r}
# 60
gc()
stm_60h1 <- metadata %>% stm(holdout1$documents, holdout1$vocab, K = 60, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout1$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_60h1, file = "results/stm_60h1.zip", compress = "xz", compression_level = 1)

# 70
gc()
stm_70h1 <- metadata %>% stm(holdout1$documents, holdout1$vocab, K = 70, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout1$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_70h1, file = "results/stm_70h1.zip", compress = "xz", compression_level = 1)

# 80
gc()
stm_80h1 <- metadata %>% stm(holdout1$documents, holdout1$vocab, K = 80, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout1$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_80h1, file = "results/stm_80h1.zip", compress = "xz", compression_level = 1)

# 90
gc()
stm_90h1 <- metadata %>% stm(holdout1$documents, holdout1$vocab, K = 90, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout1$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_90h1, file = "results/stm_90h1.zip", compress = "xz", compression_level = 1)

# 100
gc()
stm_100h1 <- metadata %>% stm(holdout1$documents, holdout1$vocab, K = 100, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout1$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_100h1, file = "results/stm_100h1.zip", compress = "xz", compression_level = 1)

# 110
gc()
stm_110h1 <- metadata %>% stm(holdout1$documents, holdout1$vocab, K = 110, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout1$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_110h1, file = "results/stm_110h1.zip", compress = "xz", compression_level = 1)

# 120
gc()
stm_120h1 <- metadata %>% stm(holdout1$documents, holdout1$vocab, K = 120, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout1$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_120h1, file = "results/stm_120h1.zip", compress = "xz", compression_level = 1)
```

## Second holdout set

```{r}
# 60
gc()
stm_60h2 <- metadata %>% stm(holdout2$documents, holdout2$vocab, K = 60, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout2$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_60h2, file = "results/stm_60h2.zip", compress = "xz", compression_level = 1)

# 70
gc()
stm_70h2 <- metadata %>% stm(holdout2$documents, holdout2$vocab, K = 70, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout2$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_70h2, file = "results/stm_70h2.zip", compress = "xz", compression_level = 1)

# 80
gc()
stm_80h2 <- metadata %>% stm(holdout2$documents, holdout2$vocab, K = 80, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout2$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_80h2, file = "results/stm_80h2.zip", compress = "xz", compression_level = 1)

# 90
gc()
stm_90h2 <- metadata %>% stm(holdout2$documents, holdout2$vocab, K = 90, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout2$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_90h2, file = "results/stm_90h2.zip", compress = "xz", compression_level = 1)

# 100
gc()
stm_100h2 <- metadata %>% stm(holdout2$documents, holdout2$vocab, K = 100, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout2$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_100h2, file = "results/stm_100h2.zip", compress = "xz", compression_level = 1)

# 110
gc()
stm_110h2 <- metadata %>% stm(holdout2$documents, holdout2$vocab, K = 110, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout2$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_110h2, file = "results/stm_110h2.zip", compress = "xz", compression_level = 1)

# 120
gc()
stm_120h2 <- metadata %>% stm(holdout2$documents, holdout2$vocab, K = 120, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout2$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_120h2, file = "results/stm_120h2.zip", compress = "xz", compression_level = 1)
```

## Third holdout set

```{r}
# 60
gc()
stm_60h3 <- metadata %>% stm(holdout3$documents, holdout3$vocab, K = 60, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout3$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_60h3, file = "results/stm_60h3.zip", compress = "xz", compression_level = 1)

# 70
gc()
stm_70h3 <- metadata %>% stm(holdout3$documents, holdout3$vocab, K = 70, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout3$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_70h3, file = "results/stm_70h3.zip", compress = "xz", compression_level = 1)

# 80
gc()
stm_80h3 <- metadata %>% stm(holdout3$documents, holdout3$vocab, K = 80, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout3$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_80h3, file = "results/stm_80h3.zip", compress = "xz", compression_level = 1)

# 90
gc()
stm_90h3 <- metadata %>% stm(holdout3$documents, holdout3$vocab, K = 90, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout3$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_90h3, file = "results/stm_90h3.zip", compress = "xz", compression_level = 1)

# 100
gc()
stm_100h3 <- metadata %>% stm(holdout3$documents, holdout3$vocab, K = 100, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout3$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_100h3, file = "results/stm_100h3.zip", compress = "xz", compression_level = 1)

# 110
gc()
stm_110h3 <- metadata %>% stm(holdout3$documents, holdout3$vocab, K = 110, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout3$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_110h3, file = "results/stm_110h3.zip", compress = "xz", compression_level = 1)

# 120
gc()
stm_120h3 <- metadata %>% stm(holdout3$documents, holdout3$vocab, K = 120, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(holdout3$vocab)), emtol = 1E-4,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_120h3, file = "results/stm_120h3.zip", compress = "xz", compression_level = 1)
```

# Evaluation

```{r}
# Holdout set 1
test1 <- c(
  eval.heldout(stm_60h1, holdout1$missing)$doc.heldout,
  eval.heldout(stm_70h1, holdout1$missing)$doc.heldout,
  eval.heldout(stm_80h1, holdout1$missing)$doc.heldout,
  eval.heldout(stm_90h1, holdout1$missing)$doc.heldout,
  eval.heldout(stm_100h1, holdout1$missing)$doc.heldout,
  eval.heldout(stm_110h1, holdout1$missing)$doc.heldout,
  eval.heldout(stm_120h1, holdout1$missing)$doc.heldout
)
n <- length(holdout1$missing$index)
# Per-document log likelihood
no1 <- test1 %>% data.frame(
  K = rep(c(60, 70, 80, 90, 100, 110, 120), each = n),
  holdout = 1,
  lik = .
) %>% na.omit()

# Holdout set 2
test2 <- c(
  eval.heldout(stm_60h2, holdout2$missing)$doc.heldout,
  eval.heldout(stm_70h2, holdout2$missing)$doc.heldout,
  eval.heldout(stm_80h2, holdout2$missing)$doc.heldout,
  eval.heldout(stm_90h2, holdout2$missing)$doc.heldout,
  eval.heldout(stm_100h2, holdout2$missing)$doc.heldout,
  eval.heldout(stm_110h2, holdout2$missing)$doc.heldout,
  eval.heldout(stm_120h2, holdout2$missing)$doc.heldout
)
n <- length(holdout2$missing$index)
# Per-document log likelihood
no2 <- test2 %>% data.frame(
  K = rep(c(60, 70, 80, 90, 100, 110, 120), each = n),
  holdout = 2,
  lik = .
) %>% na.omit()

# Holdout set 3
test3 <- c(
  eval.heldout(stm_60h3, holdout3$missing)$doc.heldout,
  eval.heldout(stm_70h3, holdout3$missing)$doc.heldout,
  eval.heldout(stm_80h3, holdout3$missing)$doc.heldout,
  eval.heldout(stm_90h3, holdout3$missing)$doc.heldout,
  eval.heldout(stm_100h3, holdout3$missing)$doc.heldout,
  eval.heldout(stm_110h3, holdout3$missing)$doc.heldout,
  eval.heldout(stm_120h3, holdout3$missing)$doc.heldout
)
n <- length(holdout3$missing$index)
# Per-document log likelihood
no3 <- test3 %>% data.frame(
  K = rep(c(60, 70, 80, 90, 100, 110, 120), each = n),
  holdout = 3,
  lik = .
) %>% na.omit()

# Combine results
stm_heldout_lik <- rbind(no1, no2, no3)

# Save results
write.csv(stm_heldout_lik, file = "results/stm_heldout_lik.csv", row.names = FALSE)
```

## Timings

```{r}
# Model timing
stm_times <- data.frame(K = rep(seq(60, 120, by = 10), times = 3), heldout = rep(1:3, each = 7),
  time = c(
    stm_60h1$time, stm_70h1$time, stm_80h1$time, stm_90h1$time, stm_100h1$time, stm_110h1$time, stm_120h1$time,
    stm_60h2$time, stm_70h2$time, stm_80h2$time, stm_90h2$time, stm_100h2$time, stm_110h2$time, stm_120h2$time,
    stm_60h3$time, stm_70h3$time, stm_80h3$time, stm_90h3$time, stm_100h3$time, stm_110h3$time, stm_120h3$time
  )/60,
  converge = c(
    stm_60h1$convergence$its, stm_70h1$convergence$its, stm_80h1$convergence$its, stm_90h1$convergence$its, stm_100h1$convergence$its, stm_110h1$convergence$its, stm_120h1$convergence$its,
    stm_60h2$convergence$its, stm_70h2$convergence$its, stm_80h2$convergence$its, stm_90h2$convergence$its, stm_100h2$convergence$its, stm_110h2$convergence$its, stm_120h2$convergence$its,
    stm_60h3$convergence$its, stm_70h3$convergence$its, stm_80h3$convergence$its, stm_90h3$convergence$its, stm_100h3$convergence$its, stm_110h3$convergence$its, stm_120h3$convergence$its
  )
)

write.csv(stm_times, "results/stm_times.csv", row.names = FALSE)
```

# Plots

```{r}
stm_heldout_lik %>% group_by(K) %>% summarise(mean.lik = mean(lik), sem.lik = sd(lik)/sqrt(n), lci = mean.lik - sem.lik, uci = mean.lik + sem.lik) %>%
  ggplot(aes(x = K, y = mean.lik)) + geom_ribbon(aes(ymin = lci, ymax = uci), alpha = .2) + geom_line(color = "blue") +
  labs(y = "mean per-document holdout log likelihood", x = "")
```

```{r}
no1 %>% group_by(K) %>% summarise(mean.lik = mean(lik), sem.lik = sd(lik)/sqrt(n), lci = mean.lik - sem.lik, uci = mean.lik + sem.lik) %>%
  ggplot(aes(x = K, y = mean.lik)) + geom_ribbon(aes(ymin = lci, ymax = uci), alpha = .2) + geom_line(color = "blue") +
  labs(y = "mean per-document holdout log likelihood", x = "")
```

```{r}
no2 %>% group_by(K) %>% summarise(mean.lik = mean(lik), sem.lik = sd(lik)/sqrt(n), lci = mean.lik - sem.lik, uci = mean.lik + sem.lik) %>%
  ggplot(aes(x = K, y = mean.lik)) + geom_ribbon(aes(ymin = lci, ymax = uci), alpha = .2) + geom_line(color = "blue") +
  labs(y = "mean per-document holdout log likelihood", x = "")
```

```{r}
no3 %>% group_by(K) %>% summarise(mean.lik = mean(lik), sem.lik = sd(lik)/sqrt(n), lci = mean.lik - sem.lik, uci = mean.lik + sem.lik) %>%
  ggplot(aes(x = K, y = mean.lik)) + geom_ribbon(aes(ymin = lci, ymax = uci), alpha = .2) + geom_line(color = "blue") +
  labs(y = "mean per-document holdout log likelihood", x = "")
```

# Final model

```{r}
# 90 full data model
gc()
stm_final_90 <- metadata %>% stm(sparse$documents, sparse$vocab, K = 90, init.type = "Spectral", ngroups = 5,
  control = list(maxV = length(sparse$vocab)), emtol = 1E-5,
  prevalence = ~ s(Score) + s(Months_Since_Feb_09) + HW_Tag,
  content = metadata$HW_Tag,
  data = .)
save(stm_final_90, file = "results/stm_final_90.zip", compress = "xz", compression_level = 1)
```
